Correction to dark frame processing.
Have corrected the handling of dark frames so that the averaging
is done using a different formula of 1/layer number.  This was
pointed out as wrong here:
http://blog.starcircleacademy.com/2012/10/darkframes/
and the confirmed here:
http://www.cambridgeincolour.com/tutorials/image-averaging-noise.htm

Also updated the dark frame to be applied using subtraction not
difference as this was also incorrect.

Updated version to 1.5.